fix: broken redirects for projects
