+++ src/libraries/table.tsx
@@ -1,4 +1,3 @@
-import { handleRedirects } from '~/utils/handleRedirects.server'
 import { PersonStanding, Zap } from 'lucide-react'
 import { twMerge } from 'tailwind-merge'
 import { CogsIcon } from '~/components/icons/CogsIcon'
@@ -15,15 +14,6 @@ export const tableProject = {
   textColor: 'text-blue-600',
   defaultDocs: 'introduction',
   legacyPackages: ['react-table'],
-  handleRedirects: (href: string) => {
-    handleRedirects(
-      reactTableV7List,
-      href,
-      '/table/v7',
-      '/table/v8',
-      'from=reactTableV7',
-    )
-  },
   testimonials: [
     {
       quote:
@@ -107,35 +97,3 @@ export const tableProject = {
     },
   ],
 }
-
-// prettier-ignore
-export const reactTableV7List = [
-    {from: 'docs/api/overview',to: 'docs/overview',},
-    {from: 'docs/api/useColumnOrder',to: 'docs/api/features/column-ordering',},
-    {from: 'docs/api/useExpanded',to: 'docs/api/features/expanding',},
-    {from: 'docs/api/useFilters',to: 'docs/api/features/filters',},
-    {from: 'docs/api/useGlobalFilter',to: 'docs/api/features/filters',},
-    {from: 'docs/api/useGroupBy',to: 'docs/api/features/grouping',},
-    {from: 'docs/api/usePagination',to: 'docs/api/features/pagination',},
-    {from: 'docs/api/useResizeColumns',to: 'docs/api/features/column-sizing',},
-    {from: 'docs/api/useRowSelect',to: 'docs/api/features/row-selection',},
-    {from: 'docs/api/useSortBy',to: 'docs/api/features/sorting',},
-    {from: 'docs/api/useTable',to: 'docs/guide/tables',},
-    {from: 'docs/examples/basic',to: 'docs/framework/react/examples/basic',},
-    {from: 'docs/examples/filtering',to: 'docs/framework/react/examples/filters',},
-    {from: 'docs/examples/footers',to: 'docs/framework/react/examples/basic',},
-    {from: 'docs/examples/grouping',to: 'docs/framework/react/examples/grouping',},
-    {from: 'docs/examples/pagination-controlled',to: 'docs/framework/react/examples/pagination-controlled',},
-    {from: 'docs/examples/pagination',to: 'docs/framework/react/examples/pagination',},
-    {from: 'docs/examples/sorting',to: 'docs/framework/react/examples/sorting',},
-    {from: 'docs/examples/row-selection',to: 'docs/framework/react/examples/row-selection',},
-    {from: 'docs/examples/row-selection-with-pagination',to: 'docs/framework/react/examples/row-selection',},
-    {from: 'docs/examples/expanding',to: 'docs/framework/react/examples/expanding',},
-    {from: 'docs/examples/editable-data',to: 'docs/framework/react/examples/editable-data',},
-    {from: 'docs/examples/column-ordering',to: 'docs/framework/react/examples/column-ordering',},
-    {from: 'docs/examples/column-hiding',to: 'docs/framework/react/examples/column-visibility',},
-    {from: 'docs/examples/column-resizing',to: 'docs/framework/react/examples/column-sizing',},
-    {from: 'docs/installation',to: 'docs/installation',},
-    {from: 'docs/overview',to: 'docs/introduction',},
-    {from: 'docs/quick-start',to: 'docs/overview',},
-]
